move solarray test to t_solution_array
